Handle missing rules in difference check
Denote a difference when the rules file is missing. Otherwise, the
function will run into an error when attempting to extract the rules.

+++ functions/Test-RuleSetDifference.ps1
@@ -7,13 +7,16 @@ function Test-RuleSetDifference {
         [Parameter(Mandatory)]
         [string]$Json
     )
-    begin {
+    process {
+        if (-not (Test-Path -Path $Path)) {
+            return $true
+        }
+
         $tempPath = [System.IO.Path]::GetTempFileName()
         $tempJsonPath = "$tempPath.json"
         Rename-Item -Path $tempPath -NewName $tempJsonPath
         & jq '.rules' $Path > $tempJsonPath
-    }
-    process {
+
         if ($null -eq (Get-Command -Name 'jd' -ErrorAction SilentlyContinue)) {
             $warningMessage = 'Checking for rule-set differences requires jd. Please download it from GitHub ' +
                 '(https://github.com/josephburnett/jd) or install it with Go (go install github.com/josephburnett/jd@latest).'
@@ -25,6 +28,8 @@ function Test-RuleSetDifference {
         return $hasDifferences
     }
     end {
-        Remove-Item -Path $tempJsonPath
+        if ($tempJsonPath) {
+            Remove-Item -Path $tempJsonPath
+        }
     }
 }
